requirements.txt
inky>=2.0.0
Pillow>=10.0.0
requests>=2.31.0

